requirements.txt
Flask==3.0.3
google-api-python-client==2.136.0
google-auth==2.29.0
google-auth-httplib2==0.2.0
openai==1.40.0
python-dotenv==1.0.1
gunicorn==22.0.0
